Re-named WorkoutHistoryAdapter to be MovementHistoryAdapter. Added PastWorkoutAdapter to handle the data in a different way to create a new expandable list.
