+++ app/src/main/java/wonderful/workouts/fragments/HomeView.java
@@ -53,9 +53,9 @@ public class HomeView extends Fragment {
 
             // Navigate to home!
             // Navigation.findNavController(view).navigate(R.id.navigation_current_workout);
-            // Navigation.findNavController(view).navigate(R.id.navigation_past_workout);
+            Navigation.findNavController(view).navigate(R.id.navigation_past_workout);
             // Navigation.findNavController(view).navigate(R.id.navigation_movements);
-            Navigation.findNavController(view).navigate(R.id.navigation_movement_history);
+            // Navigation.findNavController(view).navigate(R.id.navigation_movement_history);
             // Navigation.findNavController(view).navigate(R.id.navigation_profile_page);
             // Navigation.findNavController(view).navigate(R.id.navigation_newEditMovement_page);
             // Navigation.findNavController(view).navigate(R.id.navigation_history_page);
