+++ app/src/main/java/wonderful/workouts/fragments/MovementHistoryView.java
@@ -14,7 +14,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import wonderful.workouts.R;
-import wonderful.workouts.adapters.WorkoutWithHistoryAdapter;
+import wonderful.workouts.adapters.MovementHistoryAdapter;
 import wonderful.workouts.database.entities.Movement;
 import wonderful.workouts.database.entities.Workout;
 import wonderful.workouts.database.entities.WorkoutHistory;
@@ -69,7 +69,7 @@ public class MovementHistoryView extends Fragment {
         workoutHistories.add(dummyData);
 
         // Set the ListView's adapter to our custom adapter!
-        movementHistoryListView.setAdapter(new WorkoutWithHistoryAdapter(this.getContext(), workoutHistories));
+        movementHistoryListView.setAdapter(new MovementHistoryAdapter(this.getContext(), workoutHistories));
 
         new Thread(() -> {
             // Dummy data for now!
