+++ app/src/main/res/layout/fragment_past_workout.xml
@@ -15,11 +15,11 @@
         android:layout_marginEnd="16dp"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/WorkoutName" />
+        app:layout_constraintTop_toBottomOf="@+id/text_view_past_workout_name" />
 
     <TextView
-        android:id="@+id/WorkoutName"
-        android:layout_width="55dp"
+        android:id="@+id/text_view_past_workout_name"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="16dp"
@@ -29,15 +29,16 @@
         app:layout_constraintTop_toTopOf="parent" />
 
     <TextView
-        android:id="@+id/WorkoutDuration"
-        android:layout_width="0dp"
+        android:id="@+id/text_view_past_workout_duration"
+        android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginStart="8dp"
         android:layout_marginTop="16dp"
         android:layout_marginEnd="16dp"
         android:ems="10"
+        android:textAlignment="textEnd"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toEndOf="@+id/WorkoutName"
+        app:layout_constraintStart_toEndOf="@+id/text_view_past_workout_name"
         app:layout_constraintTop_toTopOf="parent" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
